+++ src/StandardEventFactory.php
@@ -1,16 +1,19 @@
-<?php
-
-declare(strict_types = 1);
-
-/**
- * @license See LICENSE file in project root
- */
+<?php declare(strict_types = 1);
 
 namespace Cspray\Labrador\AsyncEvent;
 
 use Cspray\Labrador\Exception\InvalidTypeException;
 use Ds\Map;
 
+/**
+ * An EventFactory implementation that will fallback to constructing a StandardEvent in the case where an Event is
+ * created with a custom factory specified.
+ *
+ * If you do want to override the type of any Event created simply call the register(string, callable) method.
+ *
+ * @package Cspray\Labrador\AsyncEvent
+ * @license See LICENSE in source root
+ */
 class StandardEventFactory implements EventFactory {
 
     private $eventFactories;
@@ -19,6 +22,14 @@ class StandardEventFactory implements EventFactory {
         $this->eventFactories = new Map();
     }
 
+    /**
+     * @param string $eventName
+     * @param object $target
+     * @param array $eventData
+     * @param mixed ...$args
+     * @return Event
+     * @throws InvalidTypeException
+     */
     public function create(string $eventName, $target, array $eventData = [], ...$args) : Event {
         if (!$this->eventFactories->hasKey($eventName)) {
             return new StandardEvent($eventName, $target, $eventData);
@@ -42,6 +53,20 @@ class StandardEventFactory implements EventFactory {
         }
     }
 
+    /**
+     * The $factoryFunction will be invoked whenever an event with the given $eventName is created.
+     *
+     * The method signature for $factoryFunction should match:
+     *
+     * function(object, array, ...args) {}
+     *
+     * The object is the target for the created Event.
+     * The array is any data that should be associated to the Event.
+     * The variadic args are any constructor arguments that the Event requires OR was simply passed to the create call.
+     *
+     * @param string $eventName
+     * @param callable $factoryFunction
+     */
     public function register(string $eventName, callable $factoryFunction) {
         $this->eventFactories->put($eventName, $factoryFunction);
     }